Address Code Reviews * removed outer struct wrappers in users api response * remove making reset password token from create user * moved checking for existing user in CreateUser method * update tests
