+++ lib/services/local/users.go
@@ -105,11 +105,21 @@ func (s *IdentityService) getUsersWithSecrets() ([]services.User, error) {
 	return users, nil
 }
 
-// CreateUser creates user if it does not exist
+// CreateUser creates user if it does not exist.
 func (s *IdentityService) CreateUser(user services.User) error {
 	if err := user.Check(); err != nil {
 		return trace.Wrap(err)
 	}
+
+	// Check if user already exists.
+	_, err := s.GetUser(user.GetName(), false)
+	if !trace.IsNotFound(err) {
+		if err != nil {
+			return trace.Wrap(err)
+		}
+		return trace.AlreadyExists("user %q already registered", user.GetName())
+	}
+
 	value, err := services.GetUserMarshaler().MarshalUser(user.WithoutSecrets().(services.User))
 	if err != nil {
 		return trace.Wrap(err)
